refactor: drop the constants.ts re-export, import labels directly
Removed the export { ISSUE_TYPE_LABELS } re-export from constants.ts
and the explanatory comment/declaration order in types.ts (deliberate
simplification). Updated the three consumers that previously imported
ISSUE_TYPE_LABELS via @/lib/constants (AccessibilityValidator.tsx,
IssuesOverviewList.tsx, IssuesWrapper.tsx) to import it directly from
@/lib/issueTypeLabels instead, since the re-export they relied on no
longer exists.

Verified with cache-cleared tsc runs against all three tsconfig
projects individually (app/plugin/node), not just `tsc -b`, after an
earlier false pass here: lint-staged's auto-fix during the previous
commit silently dropped the re-export, and `tsc -b --force` reported
success anyway off a stale .tsbuildinfo even with --force, while a
fresh `tsc --noEmit -p tsconfig.app.json` immediately caught the
three now-broken imports. Confirmed the false-pass mechanism don't
trust `tsc -b --force` alone as proof after files change outside a
tracked Edit call; clear node_modules/.tmp/*.tsbuildinfo and check
each project directly when in doubt.
