+++ src/components/organisms/IssuesWrapper.tsx
@@ -2,7 +2,8 @@ import InfoPopover from "@/components/organisms/InfoPopover";
 import Recommendations from "@/components/organisms/Recommendations";
 import { Button } from "@/components/ui/button";
 import Separator from "@/components/ui/separator";
-import { ISSUE_TYPE_LABELS, MESSAGE_TYPES } from "@/lib/constants";
+import { MESSAGE_TYPES } from "@/lib/constants";
+import ISSUE_TYPE_LABELS from "@/lib/issueTypeLabels";
 import { postMessageToBackend } from "@/lib/figmaUtils";
 import { ISSUE_RECOMMENDATIONS, ISSUES_DATA_SCHEMA } from "@/lib/issuesData";
 import { IssueType, IssueX } from "@/lib/types";
