+++ src/lib/constants.ts
@@ -1,8 +1,6 @@
 import ISSUE_TYPE_LABELS from "./issueTypeLabels";
 import { IssueType } from "./types";
 
-export { ISSUE_TYPE_LABELS };
-
 export const MESSAGE_TYPES = {
   START_QUICKCHECK: "start-quickcheck",
   CANCEL_QUICKCHECK: "cancel-quickcheck",
@@ -25,8 +23,8 @@ export const ISSUES_TYPES = Object.keys(ISSUE_TYPE_LABELS) as IssueType[];
 
 export const MIN_FONT_SIZE: number = 11;
 
-export const MIN_TOUCH_TARGET_SIZE: number = 44; // Minimum touch target size (44x44)
+export const MIN_TOUCH_TARGET_SIZE: number = 44;
 
-export const MIN_TOUCH_TARGET_SPACING: number = 8; // Minimum spacing between touch targets
+export const MIN_TOUCH_TARGET_SPACING: number = 8;
 
 export const TOUCH_TARGET_KEYWORDS = ["btn", "button", "link", "touch"];
